Remove redundancies and reorganize SDXL pipe code
